# chore: migrated assets to webp format for faster website loading