--- a/public/images/index.ts
+++ b/public/images/index.ts
@@ -1,4 +1,4 @@
 import Star from "./star3.svg";
 import AboutBg from './IMG_1024-Enhanced-NR.svg'
-import LogoMain from './logo-main.png'
+import LogoMain from './logo-main.webp'
 export { Star, AboutBg, LogoMain };