--- a/public/images/ysiStory/index.js
+++ b/public/images/ysiStory/index.js
@@ -1,5 +1,5 @@
 import YsiPic1 from "./picture1.svg"
-import YsiPic2 from "./picture2.png"
+import YsiPic2 from "./picture2.webp"
 import YsiStar from "./star.svg"
 import YsiTapStar from "./tapStar.svg"